requirements.txt
python==3.8
albumentations==1.1.0
hydra-core==1.3.2
matplotlib==3.4.3
numpy==1.22.3
omegaconf==2.3.0
opencv_python_headless==4.5.5.64
osgeo==0.0.1  --> pip install GDAL
pytorch_toolbelt==0.6.2
scikit_image==0.18.3
scikit_learn==0.24.2
scipy==1.8.0
seaborn==0.11.2
torch==1.12.1 --> conda install pytorch==1.12.1 torchvision==0.13.1 torchaudio==0.12.1 cudatoolkit=11.3 -c pytorch
cudnn==8.2.1
cudatoolkit==11.3
torchvision==0.15.1
tqdm==4.62.3
